rename Game to GameInfosList
