+++ src/components/List/GameInfosList.vue
